+++ src/osu/osu_object.rs
@@ -11,7 +11,7 @@ use crate::{
 const LEGACY_LAST_TICK_OFFSET: f64 = 36.0;
 const BASE_SCORING_DISTANCE: f64 = 100.0;
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub(crate) struct OsuObject {
     pub(crate) time: f64,
     pub(crate) pos: Pos2,
@@ -19,7 +19,7 @@ pub(crate) struct OsuObject {
     pub(crate) kind: OsuObjectKind,
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub(crate) enum OsuObjectKind {
     Circle,
     Slider {
@@ -33,7 +33,7 @@ pub(crate) enum OsuObjectKind {
     },
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub(crate) struct NestedObject {
     pub(crate) pos: Pos2,
     pub(crate) time: f64,